STModelGoelMCMC: A Bayesian Space-Time Model Using MCMC

This function implements a Bayesian space-time model using MCMC for failure time data across geographical locations. It estimates parameters associated with different components of the model, including covariates, spatial dependencies, and prior distributions.

STModelGoelMCMC(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), as.matrix(sites)),
  Z = X,
  M = Z,
  prior = list(c3 = (-2 * log(0.05)/max(dist(sites))) * 0.1, d3 = 0.1, BB1 = diag(100,
    ncol(Z)), AA1 = as.matrix(rep(0, ncol(Z))), BB2 = diag(100, ncol(Z)), AA2 =
    as.matrix(rep(0, ncol(Z))), aa1 = 2.01, bb1 = 1.005, V = diag(100, ncol(X)), MM1 =
    as.matrix(rep(0, ncol(X)))),
  iteration,
  burnin
)

Arguments

data

A matrix of failure times, where each column represents a station.

sites

A matrix of geographic coordinates for the stations (e.g., longitude and latitude).

X

A matrix of covariates associated with the parameter W. Default is a column of ones and the coordinates from sites.

Z

A matrix of covariates associated with the parameter beta. Default is the same as X.

M

A matrix of covariates associated with the parameter alpha. Default is the same as Z.

prior

A list specifying the hyperparameters for the prior distributions:

c3, d3: Shape and rate parameters for the prior of phi_w.

BB1, AA1: Mean vector and covariance matrix for the prior of gamma.

BB2, AA2: Mean vector and covariance matrix for the prior of eta. Defaults are equal to BB1 and AA1.

aa1, bb1: Shape and scale parameters for the inverse-Gamma prior of sigma^2_w.

V, MM1: Covariance matrix and mean vector for the prior of Psi.

iteration

The total number of MCMC iterations.

burnin

The number of burn-in iterations to discard.

Value

A list containing the following components:

Mgama: MCMC chain for the parameter gamma.

MgamaT: A vector of zeros and ones indicating acceptance (1) or rejection (0) of gamma proposals in the Metropolis-Hastings algorithm.

Meta: MCMC chain for the parameter eta.

MetaT: A vector of zeros and ones indicating acceptance (1) or rejection (0) of eta proposals in the Metropolis-Hastings algorithm.

Mv: MCMC chain for the parameter sigma^2.

Mb: MCMC chain for the parameter phi.

MbT: A vector of zeros and ones indicating acceptance (1) or rejection (0) of phi proposals in the Metropolis-Hastings algorithm.

MW: MCMC chain for the parameter W.

MWT: A vector of zeros and ones indicating acceptance (1) or rejection (0) of W proposals in the Metropolis-Hastings algorithm.

MPsi: MCMC chain for the parameter Psi.

MPsiT: A vector of zeros and ones indicating acceptance (1) or rejection (0) of Psi proposals in the Metropolis-Hastings algorithm.