Spatiotemporal Nonhomogeneous Poisson Model with Musa Okumoto Intensity Function and Seasonal Component

Performs a Markov Chain Monte Carlo (MCMC) procedure to estimate the parameters of a spatiotemporal nonhomogeneous Poisson model with a seasonal component. This model is designed for analyzing extreme rainfall, as proposed by Fidel Ernesto Castro Morales & Daniele Torres Rodrigues.

STModelMusaOkumotoMCMCSA(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), (1/100) * sites),
  prior = list(V = diag(100, ncol(X)), M = as.matrix(rep(0, ncol(X))), c2 = 1e-05, d2 =
    0.001, c4 = 2.01, d4 = 1.005, aa1 = 2.01, bb1 = 1.005, c3 = (-2 *
    log(0.05)/max(dist(sites))) * 0.1, d3 = 0.1),
  iteration,
  burnin
)

Arguments

data

A matrix representing the occurrence times of the event of interest at each monitoring station. Each column corresponds to the occurrence times of a specific station. Dimensions: mxn, where m is the maximum number of occurrences across the stations, and n is the number of monitoring stations.

sites

A matrix with the geographic coordinates of the monitoring stations. Dimensions: nx2.

X

A list of covariates for the scale parameter of the Musa Okumoto intensity function. Defaults to a matrix combining a column of ones (intercept) and the coordinates in sites.

prior

A list of hyperparameters for the prior distributions of the model parameters:

Psi

Multivariate normal distribution parameters M (mean vector) and V (covariance matrix). Default: A1=0, B1=diag(100, ncol(X)).

alpha

Gamma distribution parameters c2 and d2. Default:c2=1e-05,d2=0.001.

sigma^2_w

Inverse Gamma distribution parameters c4 and d4. Default: c4=2.01, d4=1.005.

phi_w

Gamma distribution parameters c3 and d3. Default: c3=(-2*log(0.05)/max(dist(sites)))*0.1, d3=0.1.

delta

arcsine distribution p(delta)=k(delta(d-delta))^{-1/2}, d=100.

f

arcsine distribution p(f)=k((f-af)(bf-f))^{-1/2}, af=1/(365+10), bf=1/(365-10).

theta

arcsine distribution p(theta)=k((theta)(2pi-theta))^{-1/2}.

iteration

Number of MCMC iterations.

burnin

Number of burn-in iterations to discard before the chains converge.

Value

A list containing:

MW

Samples of parameter W obtained during the MCMC procedure (iteration - burnin).

MWT

A binary vector indicating whether each proposed value of W was accepted (1) or rejected (0). Used to compute the acceptance rate for W.

MPsi

Samples of parameter Psi obtained during the MCMC procedure (iteration - burnin).

Mv

Samples of parameter sigma^2_w obtained during the MCMC procedure (iteration - burnin).

'
Mb

Samples of parameter phi_w obtained during the MCMC procedure (iteration - burnin).

MbT

A binary vector indicating whether each proposed value of phi_w was accepted (1) or rejected (0). Used to compute the acceptance rate for phi_w.

#'
Malpha

Samples of parameter alpha obtained during the MCMC procedure (iteration - burnin).

MalphaT

A binary vector indicating whether each proposed value of alpha was accepted (1) or rejected (0). Used to compute the acceptance rate for alpha.

Mdelta

Samples of parameter delta obtained during the MCMC procedure (iteration - burnin).

MdeltaT

A binary vector indicating whether each proposed value of delta was accepted (1) or rejected (0). Used to compute the acceptance rate for delta.

Mtheta

Samples of parameter theta obtained during the MCMC procedure (iteration - burnin).

MthetaT

A binary vector indicating whether each proposed value of delta was accepted (1) or rejected (0). Used to compute the acceptance rate for theta.

Mf

Samples of parameter f obtained during the MCMC procedure (iteration - burnin).

MfT

A binary vector indicating whether each proposed value of delta was accepted (1) or rejected (0). Used to compute the acceptance rate for f.