Spatiotemporal Nonhomogeneous Poisson Model with Weibull Intensity (MCMC)

Performs a Markov Chain Monte Carlo (MCMC) procedure to estimate the parameters of a spatiotemporal nonhomogeneous Poisson model. This model is designed for analyzing extreme rainfall, as proposed by Fidel Ernesto Castro Morales & Daniele Torres Rodrigues.

STModelWeibullMCMC(
  data,
  sites,
  X = cbind(as.matrix(rep(1, ncol(data))), as.matrix(sites)),
  Z = X,
  prior = list(aa1 = 0.001, bb1 = 0.001, aa2 = 2.01, bb2 = 1.005, c1 = (-2 *
    log(0.05)/max(dist(sites))) * 0.1, d1 = 0.1, c2 = (-2 * log(0.05)/max(dist(sites))) *
    0.1, d2 = 0.1, A1 = as.matrix(rep(0, ncol(X))), B1 = diag(100, ncol(X)), A =
    as.matrix(rep(0, ncol(Z))), B = diag(100, ncol(Z))),
  iteration,
  burnin
)

Arguments

data

A matrix representing the occurrence times of the event of interest at each monitoring station. Each column corresponds to the occurrence times of a specific station. Dimensions: mxn, where m is the maximum number of occurrences across the stations, and n is the number of monitoring stations.

sites

A matrix with the geographic coordinates of the monitoring stations. Dimensions: nx2.

X

A list of covariates for the scale parameter of the Weibull intensity function. Defaults to a matrix combining a column of ones (intercept) and the coordinates in sites.

Z

A list of covariates for the shape parameter of the Weibull intensity function. Defaults to the same covariates as X.

prior

A list of hyperparameters for the prior distributions of the model parameters:

sigma^2_w

Inverse Gamma distribution parameters aa1 and bb1. Default: aa1=0.001, bb1=0.001.

sigma^2_m

Inverse Gamma distribution parameters aa2 and bb2. Default: aa2=0.001, bb2=0.001.

phi_w

Gamma distribution parameters c1 and d1. Default: c1=(-2*log(0.05)/max(dist(sites)))*0.1, d1=0.1.

phi_m

Gamma distribution parameters c2 and d2. Default: c2=c1, d2=d1.

Psi

Multivariate normal distribution parameters A1 (mean vector) and B1 (covariance matrix). Default: A1=0, B1=diag(100, ncol(X)).

Beta

Multivariate normal distribution parameters A (mean vector) and B (covariance matrix). Default: A=0, B=diag(100, ncol(Z)).

iteration

Number of MCMC iterations.

burnin

Number of burn-in iterations to discard before the chains converge.

Value

A list containing:

MMj

Samples of parameter M obtained during the MCMC procedure (iteration - burnin).

MMT

A binary vector indicating whether each proposed value of M was accepted (1) or rejected (0). Used to compute the acceptance rate for M.

MW

Samples of parameter W obtained during the MCMC procedure (iteration - burnin).

MWT

A binary vector indicating whether each proposed value of W was accepted (1) or rejected (0). Used to compute the acceptance rate for W.

MPsi

Samples of parameter Psi obtained during the MCMC procedure (iteration - burnin).

MBeta

Samples of parameter Beta obtained during the MCMC procedure (iteration - burnin).

Mvw

Samples of parameter sigma^2_w obtained during the MCMC procedure (iteration - burnin).

Mbw

Samples of parameter phi_w obtained during the MCMC procedure (iteration - burnin).

MbwT

A binary vector indicating whether each proposed value of phi_w was accepted (1) or rejected (0). Used to compute the acceptance rate for phi_w.

Mvm

Samples of parameter sigma^2_m obtained during the MCMC procedure (iteration - burnin).

Mbm

Samples of parameter phi_m obtained during the MCMC procedure (iteration - burnin).

MbmT

A binary vector indicating whether each proposed value of phi_m was accepted (1) or rejected (0). Used to compute the acceptance rate for phi_m.