Compute Interpolated Mean Surface for Spatiotemporal Model

This function interpolates the mean values on a grid of points (DNO) for a spatiotemporal nonhomogeneous Poisson model with Goel intensity. The function utilizes the MCMC outputs from STModelGoelMCMCSA and applies a Gaussian process-based interpolation.

compute_mean_surfaceGOELSA(
  resultsSA,
  sites,
  X,
  Z,
  M,
  DNO,
  CovXNO,
  CovZNO,
  CovMNO,
  tau
)

Arguments

resultsSA

A list containing the output from STModelGoelMCMCSA, including:

MW: Samples for parameter W.

MWT: Acceptance indicators for parameter W.

MMj: Samples for parameter M.

MMT: Acceptance indicators for parameter M.

Mvw: Samples for parameter sigma^2_w.

Mvm: Samples for parameter sigma^2_m.

MBeta: Samples for parameter Beta.

Mbw: Samples for parameter phi_w.

MbwT: Acceptance indicators for phi_w.

Mbm: Samples for parameter phi_m.

MbmT: Acceptance indicators for phi_m.

MPsi: Samples for parameter Psi.

Mdelta: Samples of parameter delta obtained during the MCMC procedure (iteration - burnin).

Mtheta: Samples of parameter theta obtained during the MCMC procedure (iteration - burnin).

Mf: Samples of parameter f obtained during the MCMC procedure (iteration - burnin).

sites

A matrix with geographic coordinates of the monitoring stations.

X

Covariates for the W parameter of the Goel intensity.

Z

Covariates for the eta parameter of the Goel intensity.

M

Covariates for the gama parameter of the Goel intensity.

DNO

A grid of points where interpolation is to be performed.

CovXNO

Covariates for the W parameter at the grid points.

CovZNO

Covariates for the eta parameter at the grid points.

CovMNO

Covariates for the gama parameter at the grid points.

tau

A vector of temporal points for which the mean surface is computed.

Value

A list: (Surface) containing the interpolated mean values at the grid points for each temporal point in tau. The first column contains the mean values at the initial time step, and subsequent columns contain the differences between consecutive time steps.