Fit and Plot Nonhomogeneous Poisson Weibull Model

This function fits a nonhomogeneous Poisson Weibull model with a seasonal component to data from a specified monitoring station, generates a plot of the observed and estimated cumulative mean functions, and returns relevant metrics.

fit_weibull_modelSA(data, resultsSA, l)

Arguments

data

A matrix of occurrence times for the event of interest. Each column corresponds to a monitoring station.

resultsSA

A list containing the output from STModelWeibullMCMCSA, which includes:

MW: Samples for parameter W.

Mbeta: Samples for parameter Beta.

Malpha: Samples for parameter Alpha.

Mdelta: Samples of parameter delta obtained during the MCMC procedure (iteration - burnin).

Mtheta: Samples of parameter theta obtained during the MCMC procedure (iteration - burnin).

Mf: Samples of parameter f obtained during the MCMC procedure (iteration - burnin).

l

An integer specifying the index of the monitoring station to analyze.

Value

A list containing:

retemp

The occurrence times of the events of interest.

medy1

The estimated cumulative mean function at the event times.

mtnp

The observed cumulative mean function at the event times.